requirements.txt
aiogram==3.10.0
crystalpayio==1.0.5
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31

